refactor: socket 관련 로직 별도 커스텀 훅으로 분리
